Fix concealed text tab alignment bug
Vim docs use concealed texts for a few types of texts: tag links (e.g.
`|:edit|`), code blocks (e.g. `` `some code` ``), and tags (e.g.
`*some_tag*`). The begin/end characters are hidden from view for
convenience, but when Vim is calculating tab-width calculation, the
concealed characters *do* count towards the tab.

The way vimhelp worked is that it first called `line.expandtabs()` to
convert the hard tabs to spaces, before removing the concealed
characters. Because of that, the alignment is no longer correct unless
you add back the characters in as whitespace. This is somewhat difficult
to do because concealed characters followed by a normal space do *not*
need to be filled back in. You only want to do that when it's followed
by a tab.

To fix this, just do not called `expandtabs()` and instead use hard
tabs. This is fine because HTML/CSS uses 8 chars for tabs by default. We
count the number of concealed characters when we encounter a tag with
conceal, and then later if we see a hard tab ("\t") on the same line, we
manually inject empty spaces (one per concealed character) before the
tab to reproduce the same number of charaters before the tab.

This is similar to the fix at Neovim's doc generation tool which had a
similar bug: https://github.com/neovim/neovim/pull/20690. In that fix,
it only injects two space if a tag is immediately followed by a hard
tab, but in this fix it's more general and accounts for tags followed by
general text followed by a hard tab (since we count the concealed
characters before we encounter the tab).

Also, add README documentation for how to generate a static site.